Adjust tests to Weaviate 1.15.2 version

+++ src/main/java/technology/semi/weaviate/client/v1/misc/model/VectorIndexConfig.java
@@ -12,7 +12,7 @@ public class VectorIndexConfig {
   Integer cleanupIntervalSeconds;
   Integer efConstruction;
   Integer maxConnections;
-  Integer vectorCacheMaxObjects;
+  Long vectorCacheMaxObjects;
   Integer ef;
   Boolean skip;
   Integer dynamicEfFactor;
